WIP: T04 -- Extract credential/auth test classes from tests/unit/cli/test_client.py
