+++ tests/unit/cli/test_client.py
@@ -1,26 +1,21 @@
 """Unit tests for the HassetteCLIClient HTTP client wrapper."""
 
 import json
-import os
 from pathlib import Path
 from typing import Any
-from unittest.mock import MagicMock, patch
 
 import httpx2 as httpx
 import pytest
 from pydantic import BaseModel
 
-import hassette.cli as cli_pkg
 from hassette.cli.client import HassetteCLIClient
 from hassette.config.config import HassetteConfig
 from hassette.config.models import WebApiConfig
 from hassette.test_utils.web_manifest_helpers import make_manifest_list_response, make_manifest_response
-from hassette.web.auth.tokens import TOKEN_FILENAME
 from hassette.web.models import AppInstanceResponse
-from tests.unit.cli.conftest import REMOTE_SERVER_URL, CLIClientFactory, capture_stderr, make_cli_config
+from tests.unit.cli.conftest import REMOTE_SERVER_URL, capture_stderr, make_cli_config
 
 HEALTH_ENDPOINT = "/api/health"
-CLI_AUTH_TOKEN_ENV = "HASSETTE__CLI__AUTH_TOKEN"
 
 # Helpers
 
@@ -471,244 +466,6 @@ class TestDebugMode:
         assert "debug" not in parsed
 
 
-# Web API bearer-token credential attachment
-#
-# The credential-resolution tests below use the shared ``make_cli_config`` helper (see
-# ``tests/unit/cli/conftest.py``), which mirrors ``test_target.py``'s cli/web_api override shape.
-# ``server_url`` (and, indirectly, a non-loopback ``host``) builds a non-loopback target for
-# the credential-scoping and TLS-warning tests — the default ``host="127.0.0.1"`` keeps the
-# existing loopback-scoped test suite green as-is.
-
-
-class TestCredentialAttachment:
-    """The CLI resolves a web API bearer token and attaches it to outgoing requests."""
-
-    def test_config_auth_token_attaches_bearer_header(self, tmp_path: Path) -> None:
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token="config-token"))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert captured_headers[0]["authorization"] == "Bearer config-token"
-
-    def test_env_var_populates_config_and_attaches_bearer_header(
-        self,
-        tmp_path: Path,
-        monkeypatch: pytest.MonkeyPatch,
-    ) -> None:
-        """The token arrives via pydantic-settings, not a hand-rolled os.environ read.
-
-        Setting only the env var (no explicit ``web_api`` override) and confirming the
-        header is attached proves ``config.web_api.auth_token`` was actually populated by
-        HassetteConfig's normal settings resolution. This means, unlike its sibling tests, it
-        can't use the hermetic ``make_cli_config`` factory — see "Credential tests: prefer the
-        hermetic factory" in this directory's CLAUDE.md for why every ambient ``HASSETTE__*``
-        var is cleared below, not just the credential-precedence ones: an ambient
-        ``HASSETTE__CLI__SERVER_URL`` or ``HASSETTE__WEB_API__HOST`` pointed at a non-loopback
-        target would make ``resolve_cli_auth_token()`` skip the server-scoped
-        ``web_api.auth_token`` source entirely, failing this test for an unrelated reason.
-
-        Clearing env vars isn't enough on its own — a developer's repo-root ``.env`` or
-        ``hassette.toml`` could set the same problematic keys and survive the clear. This
-        drops the dotenv and TOML sources entirely (keeping ``init_settings``/``env_settings``
-        live) so only the real environment — the thing under test — can populate the config.
-        """
-        for key in list(os.environ):
-            if key.upper().startswith("HASSETTE__"):
-                monkeypatch.delenv(key, raising=False)
-        monkeypatch.setenv("HASSETTE__WEB_API__AUTH_TOKEN", "env-token")
-
-        class EnvOnlyConfig(HassetteConfig):
-            model_config = HassetteConfig.model_config.copy() | {"toml_file": None, "env_file": None}
-
-            @classmethod
-            def settings_customise_sources(cls, _settings_cls, init_settings, env_settings, **kwargs):  # pyright: ignore[reportIncompatibleMethodOverride]
-                return (init_settings, env_settings, kwargs["file_secret_settings"])
-
-        config = EnvOnlyConfig(token=None, data_dir=tmp_path)
-        assert config.web_api.auth_token is not None
-        assert config.web_api.auth_token.get_secret_value() == "env-token"
-
-        factory = CLIClientFactory(config)
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert captured_headers[0]["authorization"] == "Bearer env-token"
-
-    def test_falls_back_to_token_file_when_config_value_absent(self, tmp_path: Path) -> None:
-        (tmp_path / TOKEN_FILENAME).write_text("file-token", encoding="utf-8")
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert captured_headers[0]["authorization"] == "Bearer file-token"
-
-    def test_config_value_takes_precedence_over_token_file(self, tmp_path: Path) -> None:
-        (tmp_path / TOKEN_FILENAME).write_text("file-token", encoding="utf-8")
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token="config-token"))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert captured_headers[0]["authorization"] == "Bearer config-token"
-
-    def test_no_config_value_and_no_token_file_sends_no_authorization_header(self, tmp_path: Path) -> None:
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert "authorization" not in captured_headers[0]
-
-    def test_empty_token_file_treated_as_no_token(self, tmp_path: Path) -> None:
-        (tmp_path / TOKEN_FILENAME).write_text("", encoding="utf-8")
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert "authorization" not in captured_headers[0]
-
-    def test_missing_token_never_calls_generating_resolver(self, tmp_path: Path) -> None:
-        """No token configured, no token file — the CLI must not mint its own token.
-
-        A CLI-generated token would never match the running service's, so the correct
-        behavior is to send the request with no credential (letting the server 401 it),
-        not to silently create a value that looks like success.
-        """
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert "authorization" not in captured_headers[0]
-        assert not (tmp_path / TOKEN_FILENAME).exists()
-
-    def test_missing_token_401_gives_clear_hint(self, tmp_path: Path) -> None:
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path))
-        transport = make_transport(401, {"detail": "Unauthorized"})
-        client = factory.build(transport)
-        code, stderr = get_expecting_exit(client)
-        assert code == 1
-        assert "has hassette been started" in stderr
-
-    def test_resolved_token_401_omits_missing_token_hint(self, tmp_path: Path) -> None:
-        """A wrong-but-present token gets the plain server error, not the missing-token hint."""
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token="wrong-token"))
-        transport = make_transport(401, {"detail": "Invalid token"})
-        client = factory.build(transport)
-        _code, stderr = get_expecting_exit(client)
-        assert "has hassette been started" not in stderr
-
-    def test_empty_string_config_token_sends_no_authorization_header(self, tmp_path: Path) -> None:
-        """An empty-string config token (e.g. an unset env var interpolated by docker-compose
-        as ``HASSETTE__WEB_API__AUTH_TOKEN=""``) must be treated the same as no token at all:
-        no Authorization header is attached.
-        """
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token=""))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert "authorization" not in captured_headers[0]
-
-    def test_blank_config_token_falls_back_to_token_file(self, tmp_path: Path) -> None:
-        """A blank configured token must mirror resolve_auth_token()'s server-side handling:
-        treated as unset and falling through to the token file, not just suppressed. Otherwise
-        the CLI could resolve a different (missing) credential than the service actually
-        validates against, once the service falls back to a generated token for the same blank
-        config value.
-        """
-        (tmp_path / TOKEN_FILENAME).write_text("file-token", encoding="utf-8")
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token="  "))
-        client, captured_headers = factory.build_capturing_headers()
-        client.get(HEALTH_ENDPOINT, dict)
-        assert captured_headers[0]["authorization"] == "Bearer file-token"
-
-    def test_empty_string_config_token_401_gives_clear_hint(self, tmp_path: Path) -> None:
-        """An empty-string config token must not attach a header and must not resolve as
-        "present" — a resulting 401 should get the missing-token hint, not be treated as a
-        plain server error from a real-but-wrong credential.
-        """
-        factory = CLIClientFactory(make_cli_config(data_dir=tmp_path, web_api_auth_token=""))
-        transport = make_transport(401, {"detail": "Unauthorized"})
-        client = factory.build(transport)
-        _code, stderr = get_expecting_exit(client)
-        assert "has hassette been started" in stderr
-
-
-# No literal --token CLI argument for the web API credential
-
-
-class TestNoLiteralWebApiTokenArgument:
-    def test_only_run_command_defines_a_token_flag(self) -> None:
-        """The only ``--token``-shaped flag anywhere in the CLI is run.py's HA token flag.
-
-        No literal token argument exists for the web API bearer credential — it is
-        resolved exclusively from config/env/file (see ``resolve_cli_auth_token`` in
-        ``hassette/cli/target.py``), never accepted as a bare CLI argument (shell-history/
-        ``ps`` exposure risk).
-        """
-        cli_dir = Path(cli_pkg.__file__).parent
-        matches: list[str] = []
-        for path in cli_dir.rglob("*.py"):
-            text = path.read_text(encoding="utf-8")
-            if '"--token"' in text or "'--token'" in text:
-                matches.append(str(path.relative_to(cli_dir)))
-
-        assert matches == [str(Path("commands") / "run.py")], (
-            f"Unexpected --token flag definitions: {matches}. "
-            "The web API auth token must never be a literal CLI argument."
-        )
-
-
-# TLS verification (--no-verify-ssl / cli.verify_ssl)
-
-
-class TestVerifySslPassthrough:
-    def test_verify_true_by_default(self) -> None:
-        config = _make_host_port_config()
-        with patch("hassette.cli.client.httpx.Client") as mock_client_cls:
-            mock_client_cls.return_value = MagicMock()
-            HassetteCLIClient(config, json_mode=False)
-        assert mock_client_cls.call_args.kwargs["verify"] is True
-
-    def test_verify_false_from_config(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        with patch("hassette.cli.client.httpx.Client") as mock_client_cls:
-            mock_client_cls.return_value = MagicMock()
-            HassetteCLIClient(config, json_mode=False)
-        assert mock_client_cls.call_args.kwargs["verify"] is False
-
-
-# Non-loopback targets: fail open, not fast
-
-
-class TestRequestIssuedDespiteNoCredential:
-    def test_non_loopback_no_credential_still_issues_request(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL)
-        factory = CLIClientFactory(config)
-        client, captured_headers = factory.build_capturing_headers()
-        result = client.get(HEALTH_ENDPOINT, dict)
-        assert result == {}
-        assert len(captured_headers) == 1
-        assert "authorization" not in captured_headers[0]
-
-
-# Non-loopback 401: remedies split by where they apply
-
-
-class TestNonLoopback401Message:
-    def test_401_names_local_and_remote_remedies(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL)
-        transport = make_transport(401, {"detail": "Unauthorized"})
-        client = HassetteCLIClient(config, json_mode=False, transport=transport)
-        code, stderr = get_expecting_exit(client)
-        assert code == 1
-        assert "--token-file" in stderr
-        assert "cli.token_file" in stderr
-        assert CLI_AUTH_TOKEN_ENV in stderr
-        assert "trusted_proxies" in stderr
-        assert "on the remote instance" in stderr
-        assert "has hassette been started" not in stderr
-
-    def test_401_with_resolved_credential_omits_the_new_hint(self, tmp_path: Path) -> None:
-        """A wrong-but-present credential for a remote target gets the plain server error,
-        not the suppressed-credential hint — mirrors the existing loopback equivalent.
-        """
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_auth_token="wrong-token")
-        transport = make_transport(401, {"detail": "Invalid token"})
-        client = HassetteCLIClient(config, json_mode=False, transport=transport)
-        _code, stderr = get_expecting_exit(client)
-        assert "trusted_proxies" not in stderr
-
-
 # 3xx redirect responses: likely forward-auth login page
 
 
@@ -786,66 +543,3 @@ class TestTargetEcho:
         client = HassetteCLIClient(config, json_mode=True, transport=transport)
         parsed = get_json_error(client, capsys)
         assert "target" not in parsed
-
-
-# TLS-verification warning: config-sourced only, not the explicit flag
-
-
-class TestVerifySslWarning:
-    def test_config_sourced_insecure_warns(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        stderr = stderr_for_successful_get(config)
-        assert "TLS verification is disabled" in stderr
-
-    def test_flag_sourced_insecure_does_not_warn(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        stderr = stderr_for_successful_get(config, verify_ssl_flag=False, server_url_flag=None)
-        assert "TLS verification is disabled" not in stderr
-
-    def test_loopback_insecure_config_still_warns(self, tmp_path: Path) -> None:
-        """The warning is about a config-vs-flag distinction, not a loopback gate — it applies
-        regardless of whether the resulting target happens to be loopback.
-        """
-        config = make_cli_config(data_dir=tmp_path, cli_verify_ssl=False)
-        stderr = stderr_for_successful_get(config)
-        assert "TLS verification is disabled" in stderr
-
-    def test_config_sourced_insecure_warns_json_mode_error_path(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        transport = make_transport(500, {"detail": "boom"})
-        client = HassetteCLIClient(config, json_mode=True, transport=transport)
-        parsed = get_json_error(client, capsys)
-        assert parsed["tls_verified"] is False
-
-    def test_flag_sourced_insecure_omits_tls_verified_json_mode_error_path(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        transport = make_transport(500, {"detail": "boom"})
-        client = HassetteCLIClient(
-            config, json_mode=True, transport=transport, verify_ssl_flag=False, server_url_flag=None
-        )
-        parsed = get_json_error(client, capsys)
-        assert "tls_verified" not in parsed
-
-    def test_config_sourced_insecure_warns_on_network_error_path(self, tmp_path: Path) -> None:
-        """Mirrors the HTTP-error TLS-warning path: a connection error is a common failure
-        mode, and an operator relying on ``cli.verify_ssl = false`` must be warned there too,
-        not only when the failure happens to be an HTTP error response.
-        """
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        transport = make_transport(raise_exc=httpx.ConnectError)
-        client = HassetteCLIClient(config, json_mode=False, transport=transport)
-        _code, stderr = get_expecting_exit(client)
-        assert "TLS verification is disabled" in stderr
-
-    def test_flag_sourced_insecure_does_not_warn_on_network_error_path(self, tmp_path: Path) -> None:
-        config = make_cli_config(data_dir=tmp_path, cli_server_url=REMOTE_SERVER_URL, cli_verify_ssl=False)
-        transport = make_transport(raise_exc=httpx.ConnectError)
-        client = HassetteCLIClient(
-            config, json_mode=False, transport=transport, verify_ssl_flag=False, server_url_flag=None
-        )
-        _code, stderr = get_expecting_exit(client)
-        assert "TLS verification is disabled" not in stderr
